pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-tuner"
version = "0.1.0"
description = "Qwen3.5-4B SFT -> RL pipeline on Tinker with Modal serving"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "tinker==0.15.0",
  "tinker-cookbook==0.1.0",
  "python-dotenv>=1.0.0",
  "requests>=2.32.0",
  "certifi>=2024.0.0"
]

[project.optional-dependencies]
modal = [
  "modal>=1.3.0",
  "openai>=2.28.0,<3"
]

[project.scripts]
llm-tuner = "llm_tuner.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
